Implementace jádra systému (autor, recenzent, redaktor)

+++ src/main/java/cz/falcon9/redact/backend/services/impl/AdminServiceImpl.java
@@ -1,16 +1,20 @@
-package cz.falcon9.redact.backend.services;
+package cz.falcon9.redact.backend.services.impl;
 
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
-import cz.falcon9.redact.backend.data.models.User;
+import cz.falcon9.redact.backend.data.models.auth.User;
 import cz.falcon9.redact.backend.exceptions.ArgumentNotFoundException;
 import cz.falcon9.redact.backend.exceptions.InvalidArgumentException;
 import cz.falcon9.redact.backend.repositories.UserRepository;
+import cz.falcon9.redact.backend.services.AdminService;
 
 @Service
 @Secured("ROLE_ADMIN")
@@ -41,6 +45,22 @@ public class AdminServiceImpl implements AdminService {
     		throw new InvalidArgumentException("Cannot insert user with blank name!");
     	}
     	
+        // prevent admin overwrite
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        
+        if (!(authentication instanceof AnonymousAuthenticationToken)) {
+            String currentUserName = authentication.getName();
+            
+            if (currentUserName == user.getUserName()) {
+                throw new InvalidArgumentException("Admin cannot overwrite himself!");
+            }
+        }
+        
+        // prevent admin overwrite
+        if (user.getUserName() == "admin") {
+            throw new InvalidArgumentException("Admin account cannot be overwrited!");
+        }
+    	
         /*if (userRepository.existsById(user.getUserName())) {
             throw new InvalidArgumentException(String.format("User %s already exists!", user.getUserName()));
         }*/
@@ -52,6 +72,22 @@ public class AdminServiceImpl implements AdminService {
     public void deleteUser(String userName) {
         if (userName.isEmpty()) {
             throw new InvalidArgumentException("UserId is empty!");
+        }
+        
+        // prevent self deletion
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        
+        if (!(authentication instanceof AnonymousAuthenticationToken)) {
+            String currentUserName = authentication.getName();
+            
+            if (currentUserName == userName) {
+                throw new InvalidArgumentException("Admin cannot delete himself!");
+            }
+        }
+        
+        // prevent admin deletion
+        if (userName == "admin") {
+            throw new InvalidArgumentException("Admin account cannot be deleted!");
         }
         
         userRepository.deleteById(userName);
